Update to mega20200410
Minor changes to make the plug-in compatible with mega-20200410. Checked and both sending and receiving works from my side. Removed SYNC1 option as this might interfere with receiving (slows code down): better to change the value of this byte yourself in the Itho-lib and compile yourself.
Firmware-old is the previous version based on a 201909.. mega.